Do not error when trying to delete nonexistent temp file (#2434)

+++ src/Files/LocalTemporaryFile.php
@@ -40,6 +40,10 @@ class LocalTemporaryFile extends TemporaryFile
      */
     public function delete(): bool
     {
+        if (@unlink($this->filePath) || !$this->exists()) {
+            return true;
+        }
+
         return unlink($this->filePath);
     }
